Creates platform specific config and main modules
This commit separates platform specific config and plugin serving
code from the generic one. It will aid development of the Windows
plugin.

* Extracts unix sock service from main.go into main_linux.go.
* Moves linux specific config from config.go into config_linux.go.
* Renames config_test.go to config_linux_test.go.
